Create a package for nerdctl image convert
